SVM/requirements.txt
cvxopt==1.2.5
numpy==1.19.2
Pillow==8.0.0
